requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
streamlit>=1.29.0
openai>=1.6.1
langchain-google-genai>=0.0.6
langchain-google-vertexai>=0.0.6
google-cloud-aiplatform>=1.49.0
langchain-anthropic>=0.1.0 